Split usePackage hook into useGrantedPackages and useUploadedPackages
