+++ src/features/PackageDistribution/useUploadedPackages.tsx
@@ -1,6 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import useAPIQuery, { QueryResult } from "../api/useAPIQuery";
-import useAuth from "../auth/useAuth";
+import useAPIQuery from "../../api/useAPIQuery";
+import useAuth from "../../auth/useAuth";
 
 export type UploadedPackage = {
   package_id: string;
@@ -13,33 +13,17 @@ export type UploadedPackage = {
   shared_users?: { user_id: string; username: string }[];
 };
 
-export type GrantedPackage = {
-  created_at: string;
-  description: string;
-  package_id: string;
-  package_name: string;
-  public: boolean;
-  size: string;
-  version: string;
-};
-
-export default function usePackages() {
+export default function useUploadedPackages() {
   const queryClient = useQueryClient();
   const { query, mutation } = useAPIQuery();
   const { user } = useAuth();
 
-  const uploadedPackages = useQuery({
+  const { data, isLoading, error } = useQuery({
     queryKey: ["uploaded-packages", user?.username],
     queryFn: () =>
       query<UploadedPackage[]>(
         `/uploaded_packages_list/${user?.username}`, // TODO: This shouldn't be necessary, the API should identify the user from the token
       ),
-  });
-
-  const grantedPackages = useQuery({
-    queryKey: ["granted-packages", user?.username],
-    queryFn: () =>
-      query<GrantedPackage[]>(`/shared_packages_list/${user?.username}`),
   });
 
   const accessMutation = useMutation({
@@ -84,16 +68,9 @@ export default function usePackages() {
   };
 
   return {
-    uploadedPackages: {
-      data: uploadedPackages.data,
-      isLoading: uploadedPackages.isLoading,
-      isError: uploadedPackages.isError,
-    } as QueryResult<UploadedPackage[]>,
-    grantedPackages: {
-      data: grantedPackages.data,
-      isLoading: grantedPackages.isLoading,
-      isError: grantedPackages.isError,
-    } as QueryResult<GrantedPackage[]>,
+    packages: data,
+    isLoading,
+    error,
     grantAccess,
     revokeAccess,
   };
